Safari-611.1.10/LayoutTests/platform/gtk/accessibility/aria-inherits-presentational-expected.txt

item 0
item 1
item 2	item 3
End of test
This tests the hierarchy for required child elements of presentational parents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXWebArea
  AXRole: AXSection AXValue: item 0
  AXRole: AXSection AXValue: item 1
  AXRole: AXSection AXValue: item 2
  AXRole: AXSection AXValue: item 3
PASS successfullyParsed is true

TEST COMPLETE
